+++ wezterm-gui/src/termwindow/mod.rs
@@ -2032,7 +2032,9 @@ impl TermWindow {
             SendKey(key) => {
                 use keyevent::{window_mods_to_termwiz_mods, Key};
                 let mods = window_mods_to_termwiz_mods(key.mods);
-                if let Key::Code(key) = self.win_key_code_to_termwiz_key_code(&key.key) {
+                if let Key::Code(key) = self.win_key_code_to_termwiz_key_code(
+                    &key.key.resolve(self.config.key_map_preference),
+                ) {
                     pane.key_down(key, mods)?;
                 }
             }